+++ engine/src/gizmos/plugin.cpp
@@ -5,8 +5,14 @@
 #include <cubos/core/io/window.hpp>
 
 #include <cubos/engine/gizmos/plugin.hpp>
-#include <cubos/engine/renderer/plugin.hpp>
-#include <cubos/engine/screen_picker/plugin.hpp>
+#include <cubos/engine/gizmos/target.hpp>
+#include <cubos/engine/render/camera/draws_to.hpp>
+#include <cubos/engine/render/camera/perspective_camera.hpp>
+#include <cubos/engine/render/camera/plugin.hpp>
+#include <cubos/engine/render/picker/picker.hpp>
+#include <cubos/engine/render/picker/plugin.hpp>
+#include <cubos/engine/render/target/plugin.hpp>
+#include <cubos/engine/render/target/target.hpp>
 #include <cubos/engine/transform/plugin.hpp>
 #include <cubos/engine/window/plugin.hpp>
 
@@ -15,8 +21,9 @@
 CUBOS_DEFINE_TAG(cubos::engine::gizmosInitTag);
 CUBOS_DEFINE_TAG(cubos::engine::gizmosInputTag);
 CUBOS_DEFINE_TAG(cubos::engine::gizmosDrawTag);
-CUBOS_DEFINE_TAG(cubos::engine::gizmosPickTag);
 
+using cubos::core::gl::Framebuffer;
+using cubos::core::gl::FramebufferDesc;
 using cubos::core::gl::Texture2D;
 
 using cubos::core::io::MouseButtonEvent;
@@ -27,184 +34,36 @@ using cubos::core::io::WindowEvent;
 
 using namespace cubos::engine;
 
-static void iterateGizmos(std::vector<std::shared_ptr<Gizmos::Gizmo>>& gizmosVector,
-                          const std::vector<std::pair<glm::mat4, BaseRenderer::Viewport>>& cameras,
-                          GizmosRenderer& gizmosRenderer, ScreenPicker& screenPicker, const DeltaTime& deltaTime)
+static void decreaseGizmoLifespans(std::vector<std::shared_ptr<Gizmos::Gizmo>>& gizmos, const DeltaTime& dt)
 {
-    std::vector<std::size_t> toRemove;
-
-    for (std::size_t i = 0; i < gizmosVector.size(); i++)
+    for (std::size_t i = 0; i < gizmos.size(); i++)
     {
-        auto& gizmo = *(gizmosVector[i]);
-        for (const auto& [mvp, viewport] : cameras)
+        if (gizmos[i]->decreaseLifespan(dt.unscaledValue))
         {
-            gizmosRenderer.renderDevice->setViewport(viewport.position.x, viewport.position.y, viewport.size.x,
-                                                     viewport.size.y);
-
-            // Color
-            gizmosRenderer.renderDevice->setShaderPipeline(gizmosRenderer.drawPipeline);
-            gizmosRenderer.renderDevice->setFramebuffer(nullptr);
-            gizmo.draw(gizmosRenderer, Gizmos::Gizmo::DrawPhase::Color, mvp);
-
-            // Id
-            gizmosRenderer.renderDevice->setShaderPipeline(gizmosRenderer.idPipeline);
-            gizmosRenderer.renderDevice->setFramebuffer(screenPicker.framebuffer());
-
-            gizmosRenderer.idPipeline->getBindingPoint("gizmo")->setConstant(gizmo.id);
-
-            gizmo.draw(gizmosRenderer, Gizmos::Gizmo::DrawPhase::Id, mvp);
-        }
-
-        if (gizmo.decreaseLifespan(deltaTime.value()))
-        {
-            toRemove.push_back(i);
+            gizmos[i] = gizmos.back();
+            gizmos.pop_back();
+            i--;
         }
     }
-    while (!toRemove.empty())
-    {
-        std::size_t i = toRemove.back();
-        toRemove.pop_back();
-        gizmosVector[i] = gizmosVector.back();
-        gizmosVector.pop_back();
-    }
-}
-
-static void splitViewport(glm::ivec2 position, glm::ivec2 size, int count, BaseRenderer::Viewport* viewports)
-{
-    if (count == 1)
-    {
-        viewports[0].position = position;
-        viewports[0].size = size;
-    }
-    else if (count >= 2)
-    {
-        glm::ivec2 splitSize;
-        glm::ivec2 splitOffset;
-
-        // Split along the largest axis.
-        if (size.x > size.y)
-        {
-            splitSize = {size.x / 2, size.y};
-            splitOffset = {size.x / 2, 0};
-        }
-        else
-        {
-            splitSize = {size.x, size.y / 2};
-            splitOffset = {0, size.y / 2};
-        }
-
-        splitViewport(position, splitSize, count / 2, viewports);
-        splitViewport(position + splitOffset, splitSize, (count + 1) / 2, &viewports[count / 2]);
-    }
-}
-
-static std::vector<std::pair<glm::mat4, BaseRenderer::Viewport>> getWorldInfo(
-    Query<const LocalToWorld&, const Camera&>& query, const ActiveCameras& activeCameras, const glm::ivec2& screenSize)
-{
-    int cameraCount = 0;
-
-    for (int i = 0; i < 4; ++i) // NOLINT(modernize-loop-convert)
-    {
-        if (!activeCameras.entities[i].isNull())
-        {
-            cameraCount += 1;
-        }
-    }
-
-    BaseRenderer::Viewport viewports[4]{};
-
-    splitViewport({0, 0}, screenSize, cameraCount, viewports);
-
-    std::vector<std::pair<glm::mat4, BaseRenderer::Viewport>> output;
-
-    int usedCount = 0;
-    for (int i = 0; i < 4; ++i) // NOLINT(modernize-loop-convert)
-    {
-        if (activeCameras.entities[i].isNull())
-        {
-            continue;
-        }
-
-        if (auto components = query.at(activeCameras.entities[i]))
-        {
-            auto& [localToWorld, camera] = *components;
-            auto v = glm::inverse(localToWorld.mat);
-            auto p = glm::perspective(glm::radians(camera.fovY),
-                                      float(viewports[usedCount].size.x) / float(viewports[usedCount].size.y),
-                                      camera.zNear, camera.zFar);
-
-            output.emplace_back(p * v, viewports[usedCount]);
-            usedCount += 1;
-        }
-    }
-
-    return output;
-}
-
-static std::vector<std::pair<glm::mat4, BaseRenderer::Viewport>> getViewInfo(const ActiveCameras& activeCameras,
-                                                                             const glm::ivec2& screenSize)
-{
-    auto v = glm::translate(glm::mat4(1.0F), glm::vec3(-1.0F, -1.0F, 0.0F));
-    auto p = glm::ortho(-0.5F, 0.5F, -0.5F, 0.5F);
-    auto vp = v * p;
-
-    int cameraCount = 0;
-
-    for (int i = 0; i < 4; ++i) // NOLINT(modernize-loop-convert)
-    {
-        if (activeCameras.entities[i].isNull())
-        {
-            continue;
-        }
-        cameraCount += 1;
-    }
-
-    BaseRenderer::Viewport viewports[4]{};
-
-    splitViewport({0, 0}, screenSize, cameraCount, viewports);
-
-    std::vector<std::pair<glm::mat4, BaseRenderer::Viewport>> output;
-
-    int usedCount = 0;
-    for (int i = 0; i < 4; ++i) // NOLINT(modernize-loop-convert)
-    {
-        if (activeCameras.entities[i].isNull())
-        {
-            continue;
-        }
-        output.emplace_back(vp, viewports[usedCount]);
-        usedCount += 1;
-    }
-
-    return output;
-}
-
-static std::vector<std::pair<glm::mat4, BaseRenderer::Viewport>> getScreenInfo(const glm::ivec2& screenSize)
-{
-    auto v = glm::translate(glm::mat4(1.0F), glm::vec3(-1.0F, -1.0F, 0.0F));
-    auto p = glm::ortho(-0.5F, 0.5F, -0.5F, 0.5F);
-    auto vp = v * p;
-
-    BaseRenderer::Viewport viewport = {{0, 0}, {screenSize[0], screenSize[1]}};
-    std::pair<glm::mat4, BaseRenderer::Viewport> pair = {vp, viewport};
-    return {pair};
 }
 
 void cubos::engine::gizmosPlugin(Cubos& cubos)
 {
     cubos.depends(windowPlugin);
     cubos.depends(transformPlugin);
-    cubos.depends(rendererPlugin);
-    cubos.depends(screenPickerPlugin);
+    cubos.depends(cameraPlugin);
+    cubos.depends(renderPickerPlugin);
+    cubos.depends(renderTargetPlugin);
 
     cubos.resource<Gizmos>();
     cubos.resource<GizmosRenderer>();
+
+    cubos.component<GizmosTarget>();
 
     cubos.startupTag(gizmosInitTag);
 
     cubos.tag(gizmosInputTag);
     cubos.tag(gizmosDrawTag);
-    cubos.tag(gizmosPickTag);
 
     cubos.startupSystem("initialize GizmosRenderer")
         .tagged(gizmosInitTag)
@@ -212,7 +71,7 @@ void cubos::engine::gizmosPlugin(Cubos& cubos)
         .call(
             [](GizmosRenderer& gizmosRenderer, const Window& window) { gizmosRenderer.init(&window->renderDevice()); });
 
-    cubos.system("process gizmos input")
+    cubos.system("process Gizmos input")
         .tagged(gizmosInputTag)
         .after(windowPollTag)
         .before(gizmosDrawTag)
@@ -242,50 +101,181 @@ void cubos::engine::gizmosPlugin(Cubos& cubos)
             }
         });
 
-    cubos.system("draw gizmos")
+    cubos.system("draw Gizmos")
         .tagged(gizmosDrawTag)
-        .tagged(screenPickerDrawTag)
-        .after(rendererDrawTag)
-        .before(windowRenderTag)
-        .call([](Gizmos& gizmos, GizmosRenderer& gizmosRenderer, ScreenPicker& screenPicker,
-                 const ActiveCameras& activeCameras, const Window& window, const DeltaTime& deltaTime,
-                 Query<const LocalToWorld&, const Camera&> query) {
-            auto screenSize = window->framebufferSize();
+        .tagged(drawToRenderPickerTag)
+        .call([](Gizmos& gizmos, GizmosRenderer& gizmosRenderer, const DeltaTime& deltaTime,
+                 Query<Entity, RenderTarget&, RenderPicker&, GizmosTarget&> targets,
+                 Query<const LocalToWorld&, const PerspectiveCamera&, const DrawsTo&> cameras) {
+            auto& rd = *gizmosRenderer.renderDevice;
+            auto orthoVP =
+                glm::translate(glm::mat4(1.0F), glm::vec3(-1.0F, -1.0F, 0.0F)) * glm::ortho(-0.5F, 0.5F, -0.5F, 0.5F);
 
-            gizmosRenderer.renderDevice->setShaderPipeline(gizmosRenderer.drawPipeline);
-            gizmosRenderer.renderDevice->setDepthStencilState(gizmosRenderer.doDepthCheckStencilState);
-            gizmosRenderer.renderDevice->clearDepth(1.0F);
+            // Draw gizmos to each of the requested targets.
+            for (auto [targetEnt, target, picker, gizmosTarget] : targets)
+            {
+                // Prepare a framebuffer for drawing to the picker texture, if necessary.
+                if (gizmosTarget.frontPicker != picker.frontTexture)
+                {
+                    std::swap(gizmosTarget.frontPicker, gizmosTarget.backPicker);
+                    std::swap(gizmosTarget.frontFramebuffer, gizmosTarget.backFramebuffer);
+                }
+                if (gizmosTarget.frontPicker != picker.frontTexture)
+                {
+                    FramebufferDesc desc{};
+                    desc.targetCount = 1;
+                    desc.targets[0].setTexture2DTarget(picker.frontTexture);
 
-            auto worldInfo = getWorldInfo(query, activeCameras, screenSize);
-            iterateGizmos(gizmos.worldGizmos, worldInfo, gizmosRenderer, screenPicker, deltaTime);
+                    gizmosTarget.frontPicker = picker.frontTexture;
+                    gizmosTarget.frontFramebuffer = rd.createFramebuffer(desc);
+                }
 
-            gizmosRenderer.renderDevice->setDepthStencilState(gizmosRenderer.noDepthCheckStencilState);
-            auto viewInfo = getViewInfo(activeCameras, screenSize);
-            iterateGizmos(gizmos.viewGizmos, viewInfo, gizmosRenderer, screenPicker, deltaTime);
+                // Prepare for color pass.
+                rd.setRasterState(nullptr);
+                rd.setBlendState(nullptr);
+                rd.setFramebuffer(target.framebuffer);
+                if (!target.cleared)
+                {
+                    rd.clearColor(0.0F, 0.0F, 0.0F, 0.0F);
+                    rd.clearDepth(1.0F);
+                    target.cleared = true;
+                }
+                rd.setShaderPipeline(gizmosRenderer.drawPipeline);
 
-            auto screenInfo = getScreenInfo(screenSize);
-            iterateGizmos(gizmos.screenGizmos, screenInfo, gizmosRenderer, screenPicker, deltaTime);
+                // Draw world and view-space gizmos.
+                for (auto [localToWorld, camera, drawsTo] : cameras.pin(1, targetEnt))
+                {
+                    // Prepare the viewport and scissor rectangles.
+                    rd.setViewport(static_cast<int>(drawsTo.viewportOffset.x * static_cast<float>(target.size.x)),
+                                   static_cast<int>(drawsTo.viewportOffset.y * static_cast<float>(target.size.y)),
+                                   static_cast<int>(drawsTo.viewportSize.x * static_cast<float>(target.size.x)),
+                                   static_cast<int>(drawsTo.viewportSize.y * static_cast<float>(target.size.y)));
+                    rd.setScissor(static_cast<int>(drawsTo.viewportOffset.x * static_cast<float>(target.size.x)),
+                                  static_cast<int>(drawsTo.viewportOffset.y * static_cast<float>(target.size.y)),
+                                  static_cast<int>(drawsTo.viewportSize.x * static_cast<float>(target.size.x)),
+                                  static_cast<int>(drawsTo.viewportSize.y * static_cast<float>(target.size.y)));
+
+                    // Prepare camera projection (TODO: fetch this matrix from a Camera component).
+                    auto worldVP = glm::perspective(glm::radians(camera.fovY),
+                                                    (drawsTo.viewportSize.x * static_cast<float>(target.size.x)) /
+                                                        (drawsTo.viewportSize.y * static_cast<float>(target.size.y)),
+                                                    camera.zNear, camera.zFar) *
+                                   glm::inverse(localToWorld.mat);
+
+                    // Draw world-space gizmos.
+                    rd.setDepthStencilState(gizmosRenderer.doDepthCheckStencilState);
+                    for (auto& gizmo : gizmos.worldGizmos)
+                    {
+                        gizmo->draw(gizmosRenderer, Gizmos::Gizmo::DrawPhase::Color, worldVP);
+                    }
+
+                    // Draw view-space gizmos.
+                    rd.setDepthStencilState(gizmosRenderer.noDepthCheckStencilState);
+                    for (auto& gizmo : gizmos.viewGizmos)
+                    {
+                        gizmo->draw(gizmosRenderer, Gizmos::Gizmo::DrawPhase::Color, orthoVP);
+                    }
+                }
+
+                // Draw screen-space gizmos.
+                rd.setViewport(0, 0, static_cast<int>(target.size.x), static_cast<int>(target.size.y));
+                rd.setScissor(0, 0, static_cast<int>(target.size.x), static_cast<int>(target.size.y));
+                for (auto& gizmo : gizmos.screenGizmos)
+                {
+                    gizmo->draw(gizmosRenderer, Gizmos::Gizmo::DrawPhase::Color, orthoVP);
+                }
+
+                // Prepare for picking pass.
+                rd.setFramebuffer(gizmosTarget.frontFramebuffer);
+                if (!picker.cleared)
+                {
+                    rd.clearTargetColor(0, UINT16_MAX, UINT16_MAX, 0, 0);
+                    picker.cleared = true;
+                }
+                rd.setShaderPipeline(gizmosRenderer.idPipeline);
+                auto* idBP = gizmosRenderer.idPipeline->getBindingPoint("gizmo");
+
+                // Draw world and view-space gizmos.
+                for (auto [localToWorld, camera, drawsTo] : cameras.pin(1, targetEnt))
+                {
+                    // Prepare the viewport and scissor rectangles.
+                    rd.setViewport(static_cast<int>(drawsTo.viewportOffset.x * static_cast<float>(target.size.x)),
+                                   static_cast<int>(drawsTo.viewportOffset.y * static_cast<float>(target.size.y)),
+                                   static_cast<int>(drawsTo.viewportSize.x * static_cast<float>(target.size.x)),
+                                   static_cast<int>(drawsTo.viewportSize.y * static_cast<float>(target.size.y)));
+                    rd.setScissor(static_cast<int>(drawsTo.viewportOffset.x * static_cast<float>(target.size.x)),
+                                  static_cast<int>(drawsTo.viewportOffset.y * static_cast<float>(target.size.y)),
+                                  static_cast<int>(drawsTo.viewportSize.x * static_cast<float>(target.size.x)),
+                                  static_cast<int>(drawsTo.viewportSize.y * static_cast<float>(target.size.y)));
+
+                    // Prepare camera projection (TODO: fetch this matrix from a Camera component).
+                    auto worldVP = glm::perspective(glm::radians(camera.fovY),
+                                                    (drawsTo.viewportSize.x * static_cast<float>(target.size.x)) /
+                                                        (drawsTo.viewportSize.y * static_cast<float>(target.size.y)),
+                                                    camera.zNear, camera.zFar) *
+                                   glm::inverse(localToWorld.mat);
+
+                    // Draw world-space gizmos.
+                    rd.setDepthStencilState(gizmosRenderer.doDepthCheckStencilState);
+                    for (auto& gizmo : gizmos.worldGizmos)
+                    {
+                        idBP->setConstant(gizmo->id);
+                        gizmo->draw(gizmosRenderer, Gizmos::Gizmo::DrawPhase::Id, worldVP);
+                    }
+
+                    // Draw view-space gizmos.
+                    rd.setDepthStencilState(gizmosRenderer.noDepthCheckStencilState);
+                    for (auto& gizmo : gizmos.viewGizmos)
+                    {
+                        idBP->setConstant(gizmo->id);
+                        gizmo->draw(gizmosRenderer, Gizmos::Gizmo::DrawPhase::Id, orthoVP);
+                    }
+                }
+
+                // Draw screen-space gizmos.
+                rd.setViewport(0, 0, static_cast<int>(target.size.x), static_cast<int>(target.size.y));
+                rd.setScissor(0, 0, static_cast<int>(target.size.x), static_cast<int>(target.size.y));
+                for (auto& gizmo : gizmos.screenGizmos)
+                {
+                    idBP->setConstant(gizmo->id);
+                    gizmo->draw(gizmosRenderer, Gizmos::Gizmo::DrawPhase::Id, orthoVP);
+                }
+            }
+
+            // Decrease lifespan of gizmos and remove dead ones.
+            decreaseGizmoLifespans(gizmos.worldGizmos, deltaTime);
+            decreaseGizmoLifespans(gizmos.viewGizmos, deltaTime);
+            decreaseGizmoLifespans(gizmos.screenGizmos, deltaTime);
         });
 
-    cubos.system("do gizmos screen picking")
-        .tagged(gizmosPickTag)
-        .after(cubos::engine::screenPickerDrawTag)
-        .call([](GizmosRenderer& gizmosRenderer, Gizmos& gizmos, const ScreenPicker& screenPicker) {
+    cubos.system("do Gizmos picking")
+        .after(drawToRenderPickerTag)
+        .call([](GizmosRenderer& gizmosRenderer, Gizmos& gizmos,
+                 Query<const RenderTarget&, const RenderPicker&> renderPickers) {
             int mouseX = gizmosRenderer.lastMousePosition.x;
             int mouseY = gizmosRenderer.lastMousePosition.y;
 
-            uint32_t id = screenPicker.at(mouseX, mouseY);
-            if (id < static_cast<uint32_t>(1 << 31))
+            // Find the render picker of the screen.
+            for (auto [target, picker] : renderPickers)
             {
-                // Not a gizmo, treat it like empty space
-                id = UINT32_MAX;
-            }
+                if (target.framebuffer == nullptr)
+                {
+                    auto id = picker.read(static_cast<unsigned int>(mouseX), static_cast<unsigned int>(mouseY));
+                    if (id < static_cast<uint32_t>(1 << 31))
+                    {
+                        // Not a gizmo, treat it like empty space
+                        id = UINT32_MAX;
+                    }
 
-            gizmos.handleInput(id, gizmosRenderer.mousePressed);
+                    gizmos.handleInput(id, gizmosRenderer.mousePressed);
 
-            if (gizmos.mLocking)
-            {
-                gizmos.setLocked(id);
+                    if (gizmos.mLocking)
+                    {
+                        gizmos.setLocked(id);
+                    }
+
+                    break;
+                }
             }
         });
 }